requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
discord.py==1.3.1
docopt==0.6.2
idna==2.8
multidict==4.7.4
pipreqs==0.4.10
python-dotenv==0.11.0
PyYAML==5.3
requests==2.22.0
urllib3==1.25.8
websockets==8.1
yarg==0.1.9
yarl==1.4.2
